Use int formatting for distance, heading

+++ Seeeduino/trackerLoraWAN/trackerLoraWAN.ino
@@ -100,11 +100,11 @@ void statusReport(void) {
 #endif
     sprintf(fmtbuf, "{\"devAddr\":\"%s\",\"frame\":[%d,%d],\"date\":\"%d/%d/%d %02d:%02d:%02d\",", devAddr, ulcntr, dlcntr, month(), day(), year(), hour(), minute(), second()); statusLine(fmtbuf);
     sprintf(fmtbuf, "\"DR\":%d,\"margin\":%d,\"LCR\":%d,\"lastLCR\":%d,\"RSSI\":%d,\"SNR\":%.1f,\"lastSNR\":%d,", currentDR, gwmargin, pendingLCR, (millis() - lastLCR) / 1000, lastRSSI, lastSNR, (millis() - lastReceived) / 1000); statusLine(fmtbuf);
-    sprintf(fmtbuf, "\"target\":%d,\"tgtdist\":%.0f,\"tgtheading\":%.0f,\"tgtage\":%d,", currentTarget, targets[currentTarget].getDistance(), targets[currentTarget].getHeading(), targets[currentTarget].getAge()); statusLine(fmtbuf);
+    sprintf(fmtbuf, "\"target\":%d,\"tgtdist\":%d,\"tgtheading\":%.0f,\"tgtage\":%d,", currentTarget, (int)targets[currentTarget].getDistance(), targets[currentTarget].getHeading(), targets[currentTarget].getAge()); statusLine(fmtbuf);
     long lat, lon;   unsigned long age;
     gps.get_position(&lat, &lon, &age);
     sprintf(fmtbuf, "\"myage\":%d,\"nsat\":%d,\"head\":%d", age / 1000, gps.satellites(),(int)imu.getHeading()); statusLine(fmtbuf);
-    //	sprintf(fmtbuf,"\"acc\":[%d,%d,%d],\"heading\":%.0f}",imu.acc_x,imu.acc_y,imu.acc_z,imu.getHeading());
+    //	sprintf(fmtbuf,"\"acc\":[%d,%d,%d],\"heading\":%d}",imu.acc_x,imu.acc_y,imu.acc_z,(int)imu.getHeading());
     //	statusLine(fmtbuf);
     lastReport = millis();
   }
